renamed `serial_port.rs` to `serial2.rs`
